+++ BetterChests/Framework/Services/AssetHandler.cs
@@ -3,11 +3,17 @@ namespace StardewMods.BetterChests.Framework.Services;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using StardewModdingAPI.Events;
+using StardewMods.BetterChests.Framework.Interfaces;
 using StardewMods.BetterChests.Framework.Models;
+using StardewMods.BetterChests.Framework.Models.StorageOptions;
 using StardewMods.Common.Interfaces;
 using StardewMods.Common.Models;
+using StardewMods.Common.Models.Events;
 using StardewMods.Common.Services;
 using StardewMods.Common.Services.Integrations.FauxCore;
+using StardewValley.GameData.BigCraftables;
+using StardewValley.GameData.Buildings;
+using StardewValley.GameData.Locations;
 
 /// <summary>Responsible for handling assets provided by this mod.</summary>
 internal sealed class AssetHandler : BaseService
@@ -16,6 +22,7 @@ internal sealed class AssetHandler : BaseService
     private readonly IGameContentHelper gameContentHelper;
     private readonly string hslTexturePath;
     private readonly Lazy<IManagedTexture> icons;
+    private readonly IModConfig modConfig;
     private readonly string tabDataPath;
     private readonly Lazy<IManagedTexture> tabs;
     private readonly string tabTexturePath;
@@ -30,6 +37,7 @@ internal sealed class AssetHandler : BaseService
     /// <param name="gameContentHelper">Dependency used for loading game assets.</param>
     /// <param name="log">Dependency used for logging debug information to the console.</param>
     /// <param name="manifest">Dependency for accessing mod manifest.</param>
+    /// <param name="modConfig">Dependency used for accessing config data.</param>
     /// <param name="modContentHelper">Dependency used for accessing mod content.</param>
     /// <param name="themeHelper">Dependency used for swapping palettes.</param>
     public AssetHandler(
@@ -38,6 +46,7 @@ internal sealed class AssetHandler : BaseService
         IGameContentHelper gameContentHelper,
         ILog log,
         IManifest manifest,
+        IModConfig modConfig,
         IModContentHelper modContentHelper,
         IThemeHelper themeHelper)
         : base(log, manifest)
@@ -45,6 +54,7 @@ internal sealed class AssetHandler : BaseService
         // Init
         this.dataHelper = dataHelper;
         this.gameContentHelper = gameContentHelper;
+        this.modConfig = modConfig;
         this.hslTexturePath = this.ModId + "/HueBar";
         this.tabDataPath = this.ModId + "/Tabs";
         this.tabTexturePath = this.ModId + "/Tabs/Texture";
@@ -60,6 +70,7 @@ internal sealed class AssetHandler : BaseService
         // Events
         eventSubscriber.Subscribe<GameLaunchedEventArgs>(this.OnGameLaunched);
         eventSubscriber.Subscribe<AssetRequestedEventArgs>(this.OnAssetRequested);
+        eventSubscriber.Subscribe<ConfigChangedEventArgs<DefaultConfig>>(this.OnConfigChanged);
     }
 
     /// <summary>Gets the hsl colors data.</summary>
@@ -104,6 +115,99 @@ internal sealed class AssetHandler : BaseService
         if (e.Name.IsEquivalentTo(this.tabDataPath))
         {
             e.LoadFrom(this.GetTabData, AssetLoadPriority.Exclusive);
+            return;
+        }
+
+        if (e.Name.IsEquivalentTo("Data/BigCraftables")
+            && this.modConfig.StorageOptions.TryGetValue("BigCraftables", out var storageTypes))
+        {
+            e.Edit(
+                asset =>
+                {
+                    var data = asset.AsDictionary<string, BigCraftableData>().Data;
+                    foreach (var (storageId, storageOptions) in storageTypes)
+                    {
+                        if (!data.TryGetValue(storageId, out var bigCraftableData))
+                        {
+                            continue;
+                        }
+
+                        bigCraftableData.CustomFields ??= new Dictionary<string, string>();
+                        var customFieldStorageOptions =
+                            new CustomFieldsStorageOptions(_ => bigCraftableData.CustomFields);
+
+                        var temporaryStorageOptions = new TemporaryStorageOptions(
+                            customFieldStorageOptions,
+                            storageOptions);
+
+                        temporaryStorageOptions.Reset();
+                        temporaryStorageOptions.Save();
+                    }
+                });
+
+            return;
+        }
+
+        if (e.Name.IsEquivalentTo("Data/Buildings")
+            && this.modConfig.StorageOptions.TryGetValue("Buildings", out storageTypes))
+        {
+            e.Edit(
+                asset =>
+                {
+                    var data = asset.AsDictionary<string, BuildingData>().Data;
+                    foreach (var (storageId, storageOptions) in storageTypes)
+                    {
+                        if (!data.TryGetValue(storageId, out var buildingData))
+                        {
+                            continue;
+                        }
+
+                        buildingData.CustomFields ??= new Dictionary<string, string>();
+                        var customFieldStorageOptions = new CustomFieldsStorageOptions(_ => buildingData.CustomFields);
+                        var temporaryStorageOptions = new TemporaryStorageOptions(
+                            customFieldStorageOptions,
+                            storageOptions);
+
+                        temporaryStorageOptions.Reset();
+                        temporaryStorageOptions.Save();
+                    }
+                });
+
+            return;
+        }
+
+        if (e.Name.IsEquivalentTo("Data/Locations")
+            && this.modConfig.StorageOptions.TryGetValue("Locations", out storageTypes))
+        {
+            e.Edit(
+                asset =>
+                {
+                    var data = asset.AsDictionary<string, LocationData>().Data;
+                    foreach (var (storageId, storageOptions) in storageTypes)
+                    {
+                        if (!data.TryGetValue(storageId, out var locationData))
+                        {
+                            continue;
+                        }
+
+                        locationData.CustomFields ??= new Dictionary<string, string>();
+                        var customFieldStorageOptions = new CustomFieldsStorageOptions(_ => locationData.CustomFields);
+                        var temporaryStorageOptions = new TemporaryStorageOptions(
+                            customFieldStorageOptions,
+                            storageOptions);
+
+                        temporaryStorageOptions.Reset();
+                        temporaryStorageOptions.Save();
+                    }
+                });
+        }
+    }
+
+    private void OnConfigChanged(ConfigChangedEventArgs<DefaultConfig> e)
+    {
+        foreach (var dataType in e.Config.StorageOptions.Keys)
+        {
+            this.gameContentHelper.InvalidateCache($"Data/{dataType}");
         }
     }