+++ BetterChests/Framework/Services/Factory/ContainerFactory.cs
@@ -10,6 +10,8 @@ using StardewMods.Common.Services.Integrations.BetterChests.Interfaces;
 using StardewMods.Common.Services.Integrations.FauxCore;
 using StardewValley.Buildings;
 using StardewValley.GameData.BigCraftables;
+using StardewValley.GameData.Buildings;
+using StardewValley.GameData.Locations;
 using StardewValley.Menus;
 using StardewValley.Objects;
 
@@ -125,7 +127,12 @@ internal sealed class ContainerFactory : BaseService
         {
             if (!this.storageOptions.TryGetValue($"(L){location.Name}", out var storageType))
             {
-                storageType = new LocationStorageOptions(() => this.modConfig.DefaultOptions, location.GetData());
+                LocationData GetData() =>
+                    DataLoader.Locations(Game1.content).TryGetValue(location.Name, out var data)
+                        ? data
+                        : new LocationData();
+
+                storageType = new LocationStorageOptions(() => this.modConfig.DefaultOptions, GetData);
                 this.storageOptions.Add($"(L){location.Name}", storageType);
             }
 
@@ -192,7 +199,10 @@ internal sealed class ContainerFactory : BaseService
 
         if (!this.storageOptions.TryGetValue($"(B){building.buildingType.Value}", out var storageType))
         {
-            storageType = new BuildingStorageOptions(() => this.modConfig.DefaultOptions, building.GetData());
+            BuildingData GetData() =>
+                Game1.buildingData.TryGetValue(building.buildingType.Value, out var data) ? data : new BuildingData();
+
+            storageType = new BuildingStorageOptions(() => this.modConfig.DefaultOptions, GetData);
             this.storageOptions.Add($"(B){building.buildingType.Value}", storageType);
         }
 
@@ -472,10 +482,10 @@ internal sealed class ContainerFactory : BaseService
 
         if (!this.storageOptions.TryGetValue(item.QualifiedItemId, out var storageOption))
         {
-            var data =
-                ItemRegistry.GetData(item.QualifiedItemId)?.RawData as BigCraftableData ?? new BigCraftableData();
+            BigCraftableData GetData() =>
+                Game1.bigCraftableData.TryGetValue(item.ItemId, out var data) ? data : new BigCraftableData();
 
-            storageOption = new BigCraftableStorageOptions(() => this.modConfig.DefaultOptions, data);
+            storageOption = new BigCraftableStorageOptions(() => this.modConfig.DefaultOptions, GetData);
             this.storageOptions.Add(item.QualifiedItemId, storageOption);
         }